requirements.txt
# Object Detection Dependencies
ultralytics>=8.0.0
opencv-python>=4.8.0
onnxruntime>=1.15.0
numpy>=1.24.0
PyYAML>=6.0
Pillow>=10.0
python-dotenv>=1.0.0

# MongoDB Atlas
pymongo>=4.6.0
dnspython>=2.4.0

# Web Dashboard
flask>=3.0.0
flask-socketio>=5.3.0
flask-cors>=4.0.0
flask-login>=0.6.3
werkzeug>=3.0.0
python-engineio>=4.8.0
python-socketio>=5.10.0

# Report Generation
reportlab>=4.0.0
matplotlib>=3.7.0

# DeepSORT Tracking
scipy>=1.9.0
scikit-learn>=1.0.0

# Optional: For improved performance on ARM devices
# onnxruntime-arm64>=1.15.0

# Development Dependencies
pytest>=7.0.0
pytest-cov>=4.0.0
